Move packages into dev-packages and mark as editable
Previously, when packages were listed in `packages` section of the
Pipfile, running `pipenv install` would copy files into the virtualenv.
This meant local development was not possible, as changes were not
reflected into the duplicate files.

General convention is to have local packages listed in `dev-packages`
sections, with the editable flag. Thus, running `pipenv install --dev`
will work as intended.

Installation of these `fluidly` packages should not be affected as the
`setup.py` files are untouched.